# server/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IFolder": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"date_created": {
						"type": "string",
						"format": "date-time"
					},
					"date_updated": {
						"type": "string",
						"format": "date-time"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"user": {
						"$ref": "#/components/schemas/IUser"
					},
					"name": {
						"type": "string"
					},
					"done": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"date_created",
					"date_updated",
					"order",
					"isDeleted",
					"user",
					"name",
					"done"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"date_created": {
						"type": "string",
						"format": "date-time"
					},
					"date_updated": {
						"type": "string",
						"format": "date-time"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"activated": {
						"type": "boolean"
					},
					"folders": {
						"items": {
							"$ref": "#/components/schemas/IFolder"
						},
						"type": "array"
					},
					"userSessions": {
						"items": {
							"$ref": "#/components/schemas/IUserSession"
						},
						"type": "array"
					},
					"defaultFolder": {
						"$ref": "#/components/schemas/Folder"
					},
					"defaultFolderId": {
						"type": "number",
						"format": "double"
					},
					"inbox": {
						"$ref": "#/components/schemas/List"
					},
					"inboxId": {
						"type": "number",
						"format": "double"
					},
					"googleUserSub": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"date_created",
					"date_updated",
					"firstName",
					"lastName",
					"email",
					"password",
					"isDeleted",
					"activated",
					"folders",
					"userSessions",
					"defaultFolder",
					"defaultFolderId",
					"inbox",
					"inboxId",
					"googleUserSub"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserSession": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"date_created": {
						"type": "string",
						"format": "date-time"
					},
					"date_updated": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/IUser"
					},
					"isRevoked": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"date_created",
					"date_updated",
					"user",
					"isRevoked"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Folder": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Unique Identifier"
					},
					"date_created": {
						"type": "string",
						"format": "date-time",
						"description": "Date of creation"
					},
					"date_updated": {
						"type": "string",
						"format": "date-time",
						"description": "Date of update"
					},
					"name": {
						"type": "string"
					},
					"done": {
						"type": "boolean"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"lists": {
						"items": {
							"$ref": "#/components/schemas/List"
						},
						"type": "array"
					},
					"listIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"date_created",
					"date_updated",
					"name",
					"done",
					"order",
					"isDeleted",
					"user",
					"userId",
					"lists",
					"listIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Unique Identifier"
					},
					"date_created": {
						"type": "string",
						"format": "date-time",
						"description": "Date of creation"
					},
					"date_updated": {
						"type": "string",
						"format": "date-time",
						"description": "Date of update"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"activated": {
						"type": "boolean"
					},
					"googleUserSub": {
						"type": "string"
					},
					"folders": {
						"items": {
							"$ref": "#/components/schemas/Folder"
						},
						"type": "array"
					},
					"userSessions": {
						"items": {
							"$ref": "#/components/schemas/UserSession"
						},
						"type": "array"
					},
					"defaultFolder": {
						"$ref": "#/components/schemas/Folder"
					},
					"inbox": {
						"$ref": "#/components/schemas/List"
					},
					"defaultFolderId": {
						"type": "number",
						"format": "double"
					},
					"inboxId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"date_created",
					"date_updated",
					"firstName",
					"lastName",
					"email",
					"password",
					"isDeleted",
					"activated",
					"googleUserSub",
					"folders",
					"userSessions",
					"defaultFolder",
					"inbox",
					"defaultFolderId",
					"inboxId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSession": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"date_created": {
						"type": "string",
						"format": "date-time"
					},
					"date_updated": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"isRevoked": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"date_created",
					"date_updated",
					"user",
					"isRevoked"
				],
				"type": "object",
				"additionalProperties": false
			},
			"List": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"date_created": {
						"type": "string",
						"format": "date-time"
					},
					"date_updated": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"folder": {
						"$ref": "#/components/schemas/Folder"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"done": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"date_created",
					"date_updated",
					"name",
					"user",
					"folder",
					"order",
					"isDeleted",
					"done"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFolderCreateProps": {
				"properties": {
					"name": {
						"type": "string"
					},
					"user": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponseModel": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLoginResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"defaultFolder": {
						"type": "number",
						"format": "double"
					},
					"inbox": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"accessToken",
					"defaultFolder",
					"inbox"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LogInRequest": {
				"properties": {
					"idToken": {
						"type": "string"
					}
				},
				"required": [
					"idToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.email-or-password_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NativeUserLoginRequest": {
				"$ref": "#/components/schemas/Pick_IUser.email-or-password_"
			},
			"Boolean": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"IUserCreateProps": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"defaultFolder": {
						"type": "number",
						"format": "double"
					},
					"inbox": {
						"type": "number",
						"format": "double"
					},
					"googleUserSub": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "to_do_list",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/folder": {
			"post": {
				"operationId": "CreateFolder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFolder"
								}
							}
						}
					}
				},
				"tags": [
					"Folders"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IFolderCreateProps"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetFolder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IFolder"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"Folders"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user/googleLogin": {
			"post": {
				"operationId": "GoogleLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserLoginResponse"
										},
										{}
									]
								}
							}
						}
					},
					"300": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LogInRequest"
							}
						}
					}
				}
			}
		},
		"/user/nativeUserLogin": {
			"post": {
				"operationId": "NativeUserLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLoginResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NativeUserLoginRequest"
							}
						}
					}
				}
			}
		},
		"/user": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/user/nativeUserSignup": {
			"post": {
				"operationId": "NativeUserSignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreateProps"
							}
						}
					}
				}
			}
		},
		"/user/validateSession": {
			"get": {
				"operationId": "ValidateCurrentSession",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean",
									"enum": [
										true
									],
									"nullable": false
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}